+++ src/core/QueryManager.ts
@@ -8,8 +8,7 @@ import {
   Deduplicator,
 } from '../transport/Deduplicator';
 
-import forOwn from 'lodash/forOwn';
-import isEqual from 'lodash/isEqual';
+import { isEqual } from '../util/isEqual';
 
 import {
   ResultTransformer,
@@ -1016,7 +1015,8 @@ export class QueryManager {
 
   private broadcastQueries() {
     const queries = this.getApolloState().queries;
-    forOwn(this.queryListeners, (listeners: QueryListener[], queryId: string) => {
+    Object.keys(this.queryListeners).forEach((queryId: string) => {
+      const listeners = this.queryListeners[queryId];
       // XXX due to an unknown race condition listeners can sometimes be undefined here.
       // this prevents a crash but doesn't solve the root cause
       // see: https://github.com/apollostack/apollo-client/issues/833