+++ src/data/storeUtils.ts
@@ -15,8 +15,6 @@ import {
   Name,
 } from 'graphql';
 
-import isObject from 'lodash/isObject';
-
 function isStringValue(value: Value): value is StringValue {
   return value.kind === 'StringValue';
 }
@@ -146,7 +144,11 @@ export interface JsonValue {
 export type StoreValue = number | string | string[] | IdValue | JsonValue | void;
 
 export function isIdValue(idObject: StoreValue): idObject is IdValue {
-  return (isObject(idObject) && (idObject as (IdValue | JsonValue)).type === 'id');
+  return (
+    idObject != null &&
+    typeof idObject === 'object' &&
+    (idObject as (IdValue | JsonValue)).type === 'id'
+  );
 }
 
 export function toIdValue(id: string, generated = false): IdValue {
@@ -158,5 +160,9 @@ export function toIdValue(id: string, generated = false): IdValue {
 }
 
 export function isJsonValue(jsonObject: StoreValue): jsonObject is JsonValue {
-  return (isObject(jsonObject) && (jsonObject as (IdValue | JsonValue)).type === 'json');
+  return (
+    jsonObject != null &&
+    typeof jsonObject === 'object' &&
+    (jsonObject as (IdValue | JsonValue)).type === 'json'
+  );
 }