+++ src/store.ts
@@ -47,7 +47,7 @@ import {
   CustomResolverMap,
 } from './data/readFromStore';
 
-import assign from 'lodash/assign';
+import { assign } from './util/assign';
 
 export interface Store {
   data: NormalizedCache;