+++ src/transport/networkInterface.ts
@@ -1,7 +1,3 @@
-import isString from 'lodash/isString';
-import assign from 'lodash/assign';
-import mapValues from 'lodash/mapValues';
-
 import 'whatwg-fetch';
 
 import {
@@ -78,9 +74,10 @@ export interface ResponseAndOptions {
 }
 
 export function printRequest(request: Request): PrintedRequest {
-  return mapValues(request, (val: any, key: any) => {
-    return key === 'query' ? print(val) : val;
-  }) as any as PrintedRequest;
+  return {
+    ...request,
+    query: print(request.query),
+  };
 }
 
 // TODO: refactor
@@ -96,12 +93,12 @@ export class HTTPFetchNetworkInterface implements NetworkInterface {
       throw new Error('A remote enpdoint is required for a network layer');
     }
 
-    if (!isString(uri)) {
+    if (typeof uri !== 'string') {
       throw new Error('Remote endpoint must be a string');
     }
 
     this._uri = uri;
-    this._opts = assign({}, opts);
+    this._opts = { ...opts };
     this._middlewares = [];
     this._afterwares = [];
   }
@@ -160,19 +157,21 @@ export class HTTPFetchNetworkInterface implements NetworkInterface {
     request,
     options,
   }: RequestAndOptions): Promise<IResponse> {
-    return fetch(this._uri, assign({}, this._opts, {
+    return fetch(this._uri, {
+      ...this._opts,
       body: JSON.stringify(printRequest(request)),
       method: 'POST',
-    }, options, {
-      headers: assign({}, {
+      ...options,
+      headers: {
         Accept: '*/*',
         'Content-Type': 'application/json',
-      }, options.headers),
-    }));
+        ...(options.headers as { [headerName: string]: string }),
+      },
+    });
   };
 
   public query(request: Request): Promise<GraphQLResult> {
-    const options = assign({}, this._opts);
+    const options = { ...this._opts };
 
     return this.applyMiddlewares({
       request,
@@ -235,7 +234,7 @@ export function createNetworkInterface(
 
   // We want to change the API in the future so that you just pass all of the options as one
   // argument, so even though the internals work with two arguments we're warning here.
-  if (isString(uriOrInterfaceOpts)) {
+  if (typeof uriOrInterfaceOpts === 'string') {
     console.warn(`Passing the URI as the first argument to createNetworkInterface is deprecated \
 as of Apollo Client 0.5. Please pass it as the "uri" property of the network interface options.`);
     opts = secondArgOpts;